+++ packages/value_class/example/lib/model/email.dart
@@ -4,7 +4,12 @@ part 'email.value.dart';
 
 @valueClass
 class Email with _$Email {
-  const factory Email(String value) = _Email;
+  @Assert(_checkNotEmpty, 'Email cannot be empty')
+  @Assert(_checkFormat, 'Email format is invalid')
+  factory Email(String value) = _Email;
 
-  const factory Email.fromJson(String value) = _Email;
+  factory Email.fromJson(String value) = _Email;
 }
+
+bool _checkNotEmpty(String value) => value.isNotEmpty;
+bool _checkFormat(String value) => value.contains(RegExp(r'^.+@\S+\.\S+$'));