+++ packages/value_class/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 1.1.0
+
+### Added
+
+- Supported `@Assert` to generate `assert(...)` statements on the generated classes:
+
+    ```dart
+    @valueClass
+    class Email with _$Email {
+      @Assert(_checkNotEmpty, 'Email cannot be empty')
+      @Assert(_checkFormat, 'Email format is invalid')
+      factory Email(String value) = _Email;
+    
+      factory Email.fromJson(String value) = _Email;
+    }
+    
+    bool _checkNotEmpty(String value) => value.isNotEmpty;
+    bool _checkFormat(String value) => value.contains(RegExp(r'^.+@\S+\.\S+$'));
+    ```
+
 ## 1.0.2
 
 ### Fixed